Add new routing to manage.js
