+++ file.js
@@ -18,20 +18,4 @@ router.get('/:file_id', async (req, res) => {
   res.render('file', data);
 })
 
-router.get('/:file_id/manage', async (req, res) => {
-  const file_id = req.params.file_id;
-  var data = {};
-  data.invited_users = await db.getInvitedUsers([file_id]);
-  data.valid_users = await db.getValidUsersForInvite([req.app.get('username'), file_id]);
-  data.file_id = file_id;
-  console.log({data});
-  res.render('manage_file', {data});
-});
-
-router.post('/:file_id/manage/invite_user', (req, res) => {
-  const from_user = req.app.get('username');
-  const to_user = req.body.invite_user;
-  const file_id = req.params.file_id;
-  db.insertNewInvite([from_user, to_user, file_id]);
-  res.redirect('back');
-});
+router.use('/:file_id/manage', require('./manage'));
